requirements.txt
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.6
flask==3.0.3
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
packaging==24.2
pillow==10.4.0
plotly==5.24.1
pydantic==2.10.5
pydantic-core==2.27.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
retrying==1.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
strawberry-graphql==0.255.0
tenacity==9.0.0
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
werkzeug==3.0.6
zipp==3.20.2
